move auth service 1 level up
